syscalls: lowered size of syscall unique ID

+++ src/include/kernel/syscalls.h
@@ -26,7 +26,7 @@ typedef struct SyscallRequest {
     bool external;          // set for syscalls that are handled in user space
     bool retry;             // for async syscalls
 
-    uint64_t requestID;     // unique random ID for user space syscalls
+    uint16_t requestID;     // unique random ID for user space syscalls
     uint64_t function;
     uint64_t params[4];
     uint64_t ret;           // return value from the kernel to the program
